+++ CHANGELOG.md
@@ -41,6 +41,7 @@
 - [BREAKING] MSRV bumped to 1.85 (#727).
 - [BREAKING] Replaced `toml` configuration with CLI (#732).
 - [BREAKING] Renamed multiple `xxx_hash` to `xxx_commitment` in RPC API (#757).
+- Added stress test for `sync-state` endpoint (#661).
 
 ### Enhancements